add transaction support

+++ src/Being/Services/ProfService.php
@@ -4,16 +4,20 @@ namespace Being\Services;
 
 class ProfService
 {
+    protected static $transaction = 0;
+
     public static function begin()
     {
-        if (function_exists('xhprof_enable')) {
+        if (self::$transaction == 0 && function_exists('xhprof_enable')) {
+            self::$transaction++;
             xhprof_enable(XHPROF_FLAGS_CPU | XHPROF_FLAGS_MEMORY);
         }
     }
 
     public static function end()
     {
-        if (function_exists('xhprof_disable')) {
+        if (self::$transaction == 1 && function_exists('xhprof_disable')) {
+            self::$transaction--;
             $data = xhprof_disable();
             $file = (ini_get('xhprof.output_dir') ? : '/tmp') . '/' . uniqid() . '.xhprof.xhprof';
             file_put_contents($file, serialize($data));
